Deleted an old model and moved TrainingData into BT module
