+++ src/lib/modules/demo_ctrl/widgets/cubits/workout_status_cubit.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:open_eqi_sports/modules/hardware/services/treadmill_control_service.dart';
-import 'package:open_eqi_sports/modules/demo_ctrl/models/workout_status.dart';
+import 'package:open_eqi_sports/modules/hardware/bt/models/treadmill_data.dart';
 
 class TrainingStatusCubit extends Cubit<TreadmillData> {
   final TreadmillControlService _treadmillControlService;
